Replace wxArrayString with std::vector<wxString>. Use std algorithms. Use correctly sized type parameters. Remove test that won't work with std::string in test/strings/strings.cpp. Add an ostream << overload for std::vector<wxString> in tests.
The non-gui tests currently have some problems. Most surround streams, but things will be migrated to std streams anyway and moved to {fmt} (and eventually std::fmt) for text formatting.

+++ include/wx/headerctrl.h
@@ -216,7 +216,7 @@ protected:
 
     // helper function for the derived classes: update the array of column
     // indices after the number of columns changed
-    void DoResizeColumnIndices(std::vector<int>& colIndices, unsigned int count);
+    void DoResizeColumnIndices(std::vector<int>& colIndices, size_t count);
 
 protected:
     // this window doesn't look nice with the border so don't use it by default
